Move LeanDataPathComponents to Common
This can be useful in so many scenarios... if only it had tests
