+++ Common/QuantConnect.csproj
@@ -446,6 +446,7 @@
     <Compile Include="Util\FuncTextWriter.cs" />
     <Compile Include="Util\IParallelRunnerWorkItem.cs" />
     <Compile Include="Util\LeanData.cs" />
+    <Compile Include="Util\LeanDataPathComponents.cs" />
     <Compile Include="Util\MarketHoursDatabaseJsonConverter.cs" />
     <Compile Include="Util\ParallelRunner.cs" />
     <Compile Include="Util\ParallelRunnerWorker.cs" />
